Addition of multiple files
