+++ Method/__init__.py
@@ -1,0 +1,4 @@
+
+
+import sys, os
+sys.path += [os.path.dirname(__file__)]
