+++ Mode/__init__.py
@@ -1,0 +1,5 @@
+
+
+import sys, os
+sys.path += [os.path.dirname(__file__)]
+
